refactor. move sample classes into example package
