+++ src/main/webapp/WEB-INF/web.xml
@@ -13,7 +13,7 @@
     </servlet>
     <servlet>
         <servlet-name>RedirectCallback</servlet-name>
-        <servlet-class>com.auth0.Auth0RedirectServlet</servlet-class>
+        <servlet-class>com.auth0.example.Auth0RedirectServlet</servlet-class>
         <init-param>
             <param-name>com.auth0.redirect_on_success</param-name>
             <param-value>/portal/home</param-value>
@@ -65,7 +65,7 @@
     <!-- Filters -->
     <filter>
         <filter-name>AuthFilter</filter-name>
-        <filter-class>com.auth0.Auth0Filter</filter-class>
+        <filter-class>com.auth0.example.Auth0Filter</filter-class>
         <init-param>
             <param-name>com.auth0.redirect_on_authentication_error</param-name>
             <param-value>/login</param-value>
